Added extra auth tests. Made auth tests work with lifespan. Added admin credentials to settings file.

+++ app/routers/auth_router.py
@@ -4,6 +4,7 @@ from fastapi import APIRouter, Depends, HTTPException, status
 from app.repositories.users_repository import UsersRepository
 from app.schemas import UserNew, AuthData, UserInDB
 from app.auth import create_access_token, get_token_user
+from app.config import settings
 
 
 auth_router = APIRouter(prefix="/auth", tags=["Authentification"])
@@ -54,5 +55,8 @@ async def get_active_user(token: str) -> UserInDB:
                     status_code=status.HTTP_202_ACCEPTED)
 async def delete_user(token: str) -> dict:
     username = get_token_user(token)
+    if username == settings.admin_username:
+        raise HTTPException(status.HTTP_403_FORBIDDEN,
+                            "Can not delete administrator user")
     await UsersRepository.delete_username(username)
     return {"message": f"Successfully deleted user {username}"}
